requirements.txt
accelerate==0.16.0
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
confection==0.1.1
cymem==2.0.7
datasets==2.10.1
dill==0.3.6
docker-pycreds==0.4.0
easydict==1.10
filelock==3.12.3
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.34
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.3
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-contrib-python==4.8.0.76
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
pathy==0.10.2
Pillow==10.0.0
preshed==3.0.8
protobuf==4.24.2
psutil==5.9.5
pyarrow==13.0.0
pydantic==1.10.12
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
responses==0.18.0
safetensors==0.3.3
sentry-sdk==1.30.0
setproctitle==1.3.2
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
thinc==8.1.12
tokenizers==0.13.3
torch==1.13.0
torchvision==0.14.0
tqdm==4.66.1
transformers==4.26.1
typer==0.7.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wandb==0.15.9
wasabi==1.1.2
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2
gradio
